Add isBeforeGo1_16 step
